+++ ChangeLog.md
@@ -3,6 +3,12 @@ Le format du fichier est basé sur [Tenez un ChangeLog](http://keepachangelog.co
 
 ## [Non Distribué]
 
+## [4.0.9] - 28-03-2019
+### Ajouter
+- Modifie le type de donnée en base des types 'select', 'sellist' et 'radio' de text en varchar(255)
+- Attention !!! Changement de 4 fonctions pour éviter les warnings: showOutputField, showOutputCustomField, showInputField, showInputCustomField => showOutputFieldAD, showOutputCustomFieldAD, showInputFieldAD, showInputCustomFieldAD
+  Les modules non modifiés en conséquances ne seront plus compatibles
+
 ## [4.0.8] - 13-03-2019
 ### Ajouter
 - Correction sur l'appel d'une hook de fetch_lines si appellé depuis une hook.
@@ -49,7 +55,8 @@ Le format du fichier est basé sur [Tenez un ChangeLog](http://keepachangelog.co
 ## [4.0.0] - 16-07-2018
 - Version initial.
 
-[Non Distribué]: http://git.open-dsi.fr/dolibarr-extension/advancedictionaries/compare/v4.0.8...HEAD
+[Non Distribué]: http://git.open-dsi.fr/dolibarr-extension/advancedictionaries/compare/v4.0.9...HEAD
+[4.0.9]: http://git.open-dsi.fr/dolibarr-extension/advancedictionaries/commits/v4.0.9
 [4.0.8]: http://git.open-dsi.fr/dolibarr-extension/advancedictionaries/commits/v4.0.8
 [4.0.7]: http://git.open-dsi.fr/dolibarr-extension/advancedictionaries/commits/v4.0.7
 [4.0.4]: http://git.open-dsi.fr/dolibarr-extension/advancedictionaries/commits/v4.0.4